module/os: widen Linux stat fields for arm64 cross-compile
syscall.Stat_t.Nlink is uint32 on linux/arm64 but uint64 on linux/amd64.
Same story for Ino and Dev on some glibc revisions. The implicit widening
worked on amd64 but broke the arm64 cross-build in goreleaser. Cast to
uint64 at the assignment site so every Linux arch lines up.

CPython: Modules/posixmodule.c:3238 os_stat_impl

+++ module/os/stat_linux.go
@@ -23,9 +23,9 @@ func statSysFields(info goos.FileInfo) (ino, dev, nlink uint64, uid, gid uint32,
 	if !ok || sys == nil {
 		return
 	}
-	ino = sys.Ino
-	dev = sys.Dev
-	nlink = sys.Nlink
+	ino = uint64(sys.Ino)
+	dev = uint64(sys.Dev)
+	nlink = uint64(sys.Nlink)
 	uid = sys.Uid
 	gid = sys.Gid
 	atime = sys.Atim.Sec
